Rename to pill_category_type
