+++ lib/domain/initial_setting/pill_category_type/card_layout.dart
